feat: add AbstractBroModule to support dependency cache inside the module

+++ sample/app.home/src/main/java/me/xx2bab/bro/sample/home/MainActivity.java
@@ -30,7 +30,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void getApi(View view) {
-        int pi = Bro.get().getApi(ISettingsApi.class).getPi();
+        double pi = Bro.get().getApi(ISettingsApi.class).getPi();
         Toast.makeText(this, pi + "", Toast.LENGTH_SHORT).show();
     }
